requirements.txt
fastapi
uvicorn
numpy==1.26.4
chromadb==1.0.9
python-jose[cryptography]
passlib
pydantic>=2.6.4
openai
aiofiles
PyPDF2
python-docx
python-multipart
SQLAlchemy
alembic
tiktoken
langchain
langchain-community
langchain-chroma
langchain-openai

